app/requirements.txt
Flask
qrcode
Pillow
reportlab
pycups

certifi==2024.7.4
configobj==5.0.9
cryptography==44.0.1
idna==3.7
jinja2==3.1.6
requests==2.32.4
setuptools==78.1.1
twisted==24.7.0rc1
urllib3==2.5.0
